fix: Settings Hub is the real entry point — profile inline, security uses tabs
- /settings now renders profile content directly (not a redirect)
- /settings/security uses SettingsLayout with tab navigation
- /profile redirects to /settings for backwards compat
- Remove unused /settings/profile redirect page
