--- frontend/app/settings/profile/page.tsx
@@ -1,12 +1,0 @@
-"use client";
-
-import { useEffect } from "react";
-import { useRouter } from "next/navigation";
-
-export default function SettingsProfileRedirect() {
-  const router = useRouter();
-  useEffect(() => {
-    router.replace("/profile");
-  }, [router]);
-  return null;
-}
